+++ gwcelery/sentry.py
@@ -8,6 +8,7 @@ from safe_netrc import netrc, NetrcParseError
 import sentry_sdk
 from sentry_sdk.integrations import celery
 
+from . import __version__ as _version
 from .util import SPHINX
 
 log = get_logger(__name__)
@@ -46,4 +47,5 @@ def configure():
     username, _, _ = auth
     dsn = urlunparse(
         (scheme, '{}@{}'.format(username, netloc), *rest))
-    sentry_sdk.init(dsn, integrations=[celery.CeleryIntegration()])
+    sentry_sdk.init(
+        dsn, integrations=[celery.CeleryIntegration()], release=_version)